requirements_temp.txt
alembic==1.16.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.13.1
fonttools==4.58.2
fsspec==2024.6.1
greenlet==3.2.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.13.1
tifffile==2025.6.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.153
ultralytics-thop==2.0.14
urllib3==2.4.0
